# docs/swagger.yaml
basePath: /api/albums/feed
definitions:
  http.Error:
    properties:
      message:
        type: string
    type: object
...
  http.artistCreateInput:
    properties:
      cover:
        type: string
      name:
        type: string
    type: object
  http.artistCreateResponse:
    properties:
      id:
        type: integer
    type: object
  http.artistDeleteResponse:
    properties:
      status:
        type: string
    type: object
  http.artistLikeResponse:
    properties:
      status:
        type: string
    type: object
...
  models.ArtistTransfer:
    properties:
      cover:
        type: string
      id:
        type: integer
      isLiked:
        type: boolean
      name:
        type: string
    type: object
...
host: localhost:4443
info:
  contact:
    email: yarik1448kuzmin@gmail.com
    name: Fluire API Support
  description: Server API for Fluire Streaming Service Application
  title: Fluire API
  version: 1.0.1
paths:
...
  /api/artists/:
    post:
      consumes:
      - application/json
      description: Create new artist by sent object
      parameters:
      - description: Track info
        in: body
        name: artist
        required: true
        schema:
          $ref: '#/definitions/http.artistCreateInput'
      produces:
      - application/json
      responses:
        "200":
          description: Artist created
          schema:
            $ref: '#/definitions/http.artistCreateResponse'
        "400":
          description: Incorrect body
          schema:
            $ref: '#/definitions/http.Error'
        "401":
          description: User unathorized
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Create Artist
      tags:
      - Artist
  /api/artists/{artistID}/:
    delete:
      description: Delete artist with chosen ID
      produces:
      - application/json
      responses:
        "200":
          description: Artist deleted
          schema:
            $ref: '#/definitions/http.artistDeleteResponse'
        "400":
          description: Incorrect body
          schema:
            $ref: '#/definitions/http.Error'
        "401":
          description: User unathorized
          schema:
            $ref: '#/definitions/http.Error'
        "403":
          description: User hasn't rights
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Delete Artist
      tags:
      - Artist
    get:
      description: Get artist with chosen ID
      produces:
      - application/json
      responses:
        "200":
          description: Artist got
          schema:
            $ref: '#/definitions/models.ArtistTransfer'
        "400":
          description: Incorrect body
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Get Artist
      tags:
      - Artist
...
  /api/artists/{artistID}/like:
    post:
      description: Set like by user to chosen artist (add to favorite)
      produces:
      - application/json
      responses:
        "200":
          description: Like set
          schema:
            $ref: '#/definitions/http.artistLikeResponse'
        "400":
          description: Client error
          schema:
            $ref: '#/definitions/http.Error'
        "401":
          description: User unathorized
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Set like
      tags:
      - Artist
...
  /api/artists/{artistID}/unlike:
    post:
      description: Remove like by user from chosen artist (remove from favorite)
      produces:
      - application/json
      responses:
        "200":
          description: Like removed
          schema:
            $ref: '#/definitions/http.artistLikeResponse'
        "400":
          description: Client error
          schema:
            $ref: '#/definitions/http.Error'
        "401":
          description: User unathorized
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Remove like
      tags:
      - Artist
  /api/artists/feed:
    get:
      description: Feed artists
      produces:
      - application/json
      responses:
        "200":
          description: Artists feed
          schema:
            items:
              $ref: '#/definitions/models.ArtistTransfer'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Artist Feed
      tags:
      - Feed
...
  /api/users/{userID}/favorite/artists:
    get:
      description: Get user's favorite artists
      produces:
      - application/json
      responses:
        "200":
          description: Artists got
          schema:
            items:
              $ref: '#/definitions/models.ArtistTransfer'
            type: array
        "400":
          description: Incorrect input
          schema:
            $ref: '#/definitions/http.Error'
        "401":
          description: Unauthorized user
          schema:
            $ref: '#/definitions/http.Error'
        "403":
          description: Forbidden user
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Favorite Artists
      tags:
      - Favorite
...
swagger: "2.0"